Fix spotless lint errors.
