+++ app/src/main/java/deakin/gopher/guardian/view/caretaker/notifications/confirmincident/ConfirmIncidentActivity.kt
@@ -12,9 +12,8 @@ import androidx.core.view.GravityCompat
 import androidx.drawerlayout.widget.DrawerLayout
 import com.google.android.material.navigation.NavigationView
 import deakin.gopher.guardian.R
-import deakin.gopher.guardian.view.FallDetection.FallAlertActivity
+import deakin.gopher.guardian.view.falldetection.FallAlertActivity
 import deakin.gopher.guardian.view.general.BaseActivity
-import deakin.gopher.guardian.view.general.Homepage4caretaker
 
 class ConfirmIncidentActivity : BaseActivity() {
     var hospitalSpinner: Spinner? = null
