+++ app/src/main/java/deakin/gopher/guardian/view/falldetection/ui/PoseGraphic.kt
@@ -1,0 +1,233 @@
+package deakin.gopher.guardian.view.falldetection.ui
+
+import android.graphics.Canvas
+import android.graphics.Color
+import android.graphics.Paint
+import com.google.mlkit.vision.pose.Pose
+import com.google.mlkit.vision.pose.PoseLandmark
+import java.lang.Math.max
+import java.lang.Math.min
+import java.util.Locale
+
+/**
+ * See original:
+ * https://github.com/googlesamples/mlkit/blob/master/android/vision-quickstart/app/src/main/java/com/google/mlkit/vision/demo/kotlin/posedetector/PoseGraphic.kt
+ */
+class PoseGraphic
+    internal constructor(
+        overlay: GraphicOverlay,
+        private val pose: Pose,
+        private val showInFrameLikelihood: Boolean,
+        private val visualizeZ: Boolean,
+        private val rescaleZForVisualization: Boolean,
+        private val poseClassification: List<String>,
+    ) : GraphicOverlay.Graphic(overlay) {
+        private var zMin = java.lang.Float.MAX_VALUE
+        private var zMax = java.lang.Float.MIN_VALUE
+        private val classificationTextPaint: Paint = Paint()
+        private val leftPaint: Paint
+        private val rightPaint: Paint
+        private val whitePaint: Paint
+
+        init {
+            classificationTextPaint.color = Color.WHITE
+            classificationTextPaint.textSize = POSE_CLASSIFICATION_TEXT_SIZE
+            classificationTextPaint.setShadowLayer(5.0f, 0f, 0f, Color.BLACK)
+
+            whitePaint = Paint()
+            whitePaint.strokeWidth = STROKE_WIDTH
+            whitePaint.color = Color.WHITE
+            whitePaint.textSize = IN_FRAME_LIKELIHOOD_TEXT_SIZE
+            leftPaint = Paint()
+            leftPaint.strokeWidth = STROKE_WIDTH
+            leftPaint.color = Color.GREEN
+            rightPaint = Paint()
+            rightPaint.strokeWidth = STROKE_WIDTH
+            rightPaint.color = Color.YELLOW
+        }
+
+        constructor(
+            overlay: GraphicOverlay,
+            pose: Pose,
+        ) : this(
+            overlay = overlay,
+            pose = pose,
+            showInFrameLikelihood = false,
+            visualizeZ = false,
+            rescaleZForVisualization = false,
+            poseClassification = emptyList(),
+        )
+
+        override fun draw(canvas: Canvas) {
+            val landmarks = pose.allPoseLandmarks
+            if (landmarks.isEmpty()) {
+                return
+            }
+
+            // Draw pose classification text.
+            val classificationX = POSE_CLASSIFICATION_TEXT_SIZE * 0.5f
+            for (i in poseClassification.indices) {
+                val classificationY =
+                    canvas.height -
+                        (POSE_CLASSIFICATION_TEXT_SIZE * 1.5f * (poseClassification.size - i).toFloat())
+                canvas.drawText(
+                    poseClassification[i],
+                    classificationX,
+                    classificationY,
+                    classificationTextPaint,
+                )
+            }
+
+            // Draw all the points
+            for (landmark in landmarks) {
+                drawPoint(canvas, landmark, whitePaint)
+                if (visualizeZ && rescaleZForVisualization) {
+                    zMin = min(zMin, landmark.position3D.z)
+                    zMax = max(zMax, landmark.position3D.z)
+                }
+            }
+
+            val nose = pose.getPoseLandmark(PoseLandmark.NOSE)
+            val lefyEyeInner = pose.getPoseLandmark(PoseLandmark.LEFT_EYE_INNER)
+            val lefyEye = pose.getPoseLandmark(PoseLandmark.LEFT_EYE)
+            val leftEyeOuter = pose.getPoseLandmark(PoseLandmark.LEFT_EYE_OUTER)
+            val rightEyeInner = pose.getPoseLandmark(PoseLandmark.RIGHT_EYE_INNER)
+            val rightEye = pose.getPoseLandmark(PoseLandmark.RIGHT_EYE)
+            val rightEyeOuter = pose.getPoseLandmark(PoseLandmark.RIGHT_EYE_OUTER)
+            val leftEar = pose.getPoseLandmark(PoseLandmark.LEFT_EAR)
+            val rightEar = pose.getPoseLandmark(PoseLandmark.RIGHT_EAR)
+            val leftMouth = pose.getPoseLandmark(PoseLandmark.LEFT_MOUTH)
+            val rightMouth = pose.getPoseLandmark(PoseLandmark.RIGHT_MOUTH)
+
+            val leftShoulder = pose.getPoseLandmark(PoseLandmark.LEFT_SHOULDER)
+            val rightShoulder = pose.getPoseLandmark(PoseLandmark.RIGHT_SHOULDER)
+            val leftElbow = pose.getPoseLandmark(PoseLandmark.LEFT_ELBOW)
+            val rightElbow = pose.getPoseLandmark(PoseLandmark.RIGHT_ELBOW)
+            val leftWrist = pose.getPoseLandmark(PoseLandmark.LEFT_WRIST)
+            val rightWrist = pose.getPoseLandmark(PoseLandmark.RIGHT_WRIST)
+            val leftHip = pose.getPoseLandmark(PoseLandmark.LEFT_HIP)
+            val rightHip = pose.getPoseLandmark(PoseLandmark.RIGHT_HIP)
+            val leftKnee = pose.getPoseLandmark(PoseLandmark.LEFT_KNEE)
+            val rightKnee = pose.getPoseLandmark(PoseLandmark.RIGHT_KNEE)
+            val leftAnkle = pose.getPoseLandmark(PoseLandmark.LEFT_ANKLE)
+            val rightAnkle = pose.getPoseLandmark(PoseLandmark.RIGHT_ANKLE)
+
+            val leftPinky = pose.getPoseLandmark(PoseLandmark.LEFT_PINKY)
+            val rightPinky = pose.getPoseLandmark(PoseLandmark.RIGHT_PINKY)
+            val leftIndex = pose.getPoseLandmark(PoseLandmark.LEFT_INDEX)
+            val rightIndex = pose.getPoseLandmark(PoseLandmark.RIGHT_INDEX)
+            val leftThumb = pose.getPoseLandmark(PoseLandmark.LEFT_THUMB)
+            val rightThumb = pose.getPoseLandmark(PoseLandmark.RIGHT_THUMB)
+            val leftHeel = pose.getPoseLandmark(PoseLandmark.LEFT_HEEL)
+            val rightHeel = pose.getPoseLandmark(PoseLandmark.RIGHT_HEEL)
+            val leftFootIndex = pose.getPoseLandmark(PoseLandmark.LEFT_FOOT_INDEX)
+            val rightFootIndex = pose.getPoseLandmark(PoseLandmark.RIGHT_FOOT_INDEX)
+
+            // Face
+            drawLine(canvas, nose, lefyEyeInner, whitePaint)
+            drawLine(canvas, lefyEyeInner, lefyEye, whitePaint)
+            drawLine(canvas, lefyEye, leftEyeOuter, whitePaint)
+            drawLine(canvas, leftEyeOuter, leftEar, whitePaint)
+            drawLine(canvas, nose, rightEyeInner, whitePaint)
+            drawLine(canvas, rightEyeInner, rightEye, whitePaint)
+            drawLine(canvas, rightEye, rightEyeOuter, whitePaint)
+            drawLine(canvas, rightEyeOuter, rightEar, whitePaint)
+            drawLine(canvas, leftMouth, rightMouth, whitePaint)
+
+            drawLine(canvas, leftShoulder, rightShoulder, whitePaint)
+            drawLine(canvas, leftHip, rightHip, whitePaint)
+
+            // Left body
+            drawLine(canvas, leftShoulder, leftElbow, leftPaint)
+            drawLine(canvas, leftElbow, leftWrist, leftPaint)
+            drawLine(canvas, leftShoulder, leftHip, leftPaint)
+            drawLine(canvas, leftHip, leftKnee, leftPaint)
+            drawLine(canvas, leftKnee, leftAnkle, leftPaint)
+            drawLine(canvas, leftWrist, leftThumb, leftPaint)
+            drawLine(canvas, leftWrist, leftPinky, leftPaint)
+            drawLine(canvas, leftWrist, leftIndex, leftPaint)
+            drawLine(canvas, leftIndex, leftPinky, leftPaint)
+            drawLine(canvas, leftAnkle, leftHeel, leftPaint)
+            drawLine(canvas, leftHeel, leftFootIndex, leftPaint)
+
+            // Right body
+            drawLine(canvas, rightShoulder, rightElbow, rightPaint)
+            drawLine(canvas, rightElbow, rightWrist, rightPaint)
+            drawLine(canvas, rightShoulder, rightHip, rightPaint)
+            drawLine(canvas, rightHip, rightKnee, rightPaint)
+            drawLine(canvas, rightKnee, rightAnkle, rightPaint)
+            drawLine(canvas, rightWrist, rightThumb, rightPaint)
+            drawLine(canvas, rightWrist, rightPinky, rightPaint)
+            drawLine(canvas, rightWrist, rightIndex, rightPaint)
+            drawLine(canvas, rightIndex, rightPinky, rightPaint)
+            drawLine(canvas, rightAnkle, rightHeel, rightPaint)
+            drawLine(canvas, rightHeel, rightFootIndex, rightPaint)
+
+            // Draw inFrameLikelihood for all points
+            if (showInFrameLikelihood) {
+                for (landmark in landmarks) {
+                    canvas.drawText(
+                        String.format(Locale.US, "%.2f", landmark.inFrameLikelihood),
+                        translateX(landmark.position.x),
+                        translateY(landmark.position.y),
+                        whitePaint,
+                    )
+                }
+            }
+        }
+
+        internal fun drawPoint(
+            canvas: Canvas,
+            landmark: PoseLandmark,
+            paint: Paint,
+        ) {
+            val point = landmark.position3D
+            updatePaintColorByZValue(
+                paint,
+                canvas,
+                visualizeZ,
+                rescaleZForVisualization,
+                point.z,
+                zMin,
+                zMax,
+            )
+            canvas.drawCircle(translateX(point.x), translateY(point.y), DOT_RADIUS, paint)
+        }
+
+        internal fun drawLine(
+            canvas: Canvas,
+            startLandmark: PoseLandmark?,
+            endLandmark: PoseLandmark?,
+            paint: Paint,
+        ) {
+            val start = startLandmark!!.position3D
+            val end = endLandmark!!.position3D
+
+            // Gets average z for the current body line
+            val avgZInImagePixel = (start.z + end.z) / 2
+            updatePaintColorByZValue(
+                paint,
+                canvas,
+                visualizeZ,
+                rescaleZForVisualization,
+                avgZInImagePixel,
+                zMin,
+                zMax,
+            )
+
+            canvas.drawLine(
+                translateX(start.x),
+                translateY(start.y),
+                translateX(end.x),
+                translateY(end.y),
+                paint,
+            )
+        }
+
+        companion object {
+            private val DOT_RADIUS = 8.0f
+            private val IN_FRAME_LIKELIHOOD_TEXT_SIZE = 30.0f
+            private val STROKE_WIDTH = 10.0f
+            private val POSE_CLASSIFICATION_TEXT_SIZE = 60.0f
+        }
+    }
