+++ app/src/main/java/deakin/gopher/guardian/view/general/MainActivity.kt
@@ -9,8 +9,6 @@ import com.google.firebase.messaging.FirebaseMessaging
 import deakin.gopher.guardian.R
 import deakin.gopher.guardian.model.login.SessionManager
 import deakin.gopher.guardian.services.EmailPasswordAuthService
-import deakin.gopher.guardian.view.FallDetection.FallAlertActivity
-import deakin.gopher.guardian.view.caretaker.notifications.confirmincident.ConfirmIncidentActivity
 
 class MainActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -18,21 +16,20 @@ class MainActivity : BaseActivity() {
         setContentView(R.layout.activity_main)
         val getStartedButton = findViewById<Button>(R.id.getStartedButton)
 
-           getStartedButton.setOnClickListener { _ -> onGetStartedClick() }
+        getStartedButton.setOnClickListener { _ -> onGetStartedClick() }
 
-            FirebaseMessaging.getInstance()
-                 .token
-                 .addOnCompleteListener { task: Task<String?> ->
-                     if (task.isSuccessful) {
-                         val token = task.result
-                         Log.d("FCM Token", token ?: "Token is null")
-                     } else {
-                         Log.w("MainActivity", "Fetching FCM registration token failed", task.exception)
-                     }
-                 }
+        FirebaseMessaging.getInstance()
+            .token
+            .addOnCompleteListener { task: Task<String?> ->
+                if (task.isSuccessful) {
+                    val token = task.result
+                    Log.d("FCM Token", token ?: "Token is null")
+                } else {
+                    Log.w("MainActivity", "Fetching FCM registration token failed", task.exception)
+                }
+            }
 
-
-        getStartedButton.setOnClickListener{
+        getStartedButton.setOnClickListener {
             startActivity(Intent(this@MainActivity, Homepage4caretaker::class.java))
         }
     }
